# Fix perf analyzer CAPI request lifecycle