--- a/src/c++/perf_analyzer/client_backend/triton_c_api/triton_loader.h
+++ b/src/c++/perf_analyzer/client_backend/triton_c_api/triton_loader.h
@@ -119,6 +119,11 @@ class TritonLoader : public tc::InferenceServerClient {
 
   static bool ModelIsLoaded() { return GetSingleton()->model_is_loaded_; }
   static bool ServerIsReady() { return GetSingleton()->server_is_ready_; }
+  static TRITONSERVER_Error* DeleteInferRequest(
+      TRITONSERVER_InferenceRequest* irequest)
+  {
+    return GetSingleton()->request_delete_fn_(irequest);
+  }
 
   // TRITONSERVER_ApiVersion
   typedef TRITONSERVER_Error* (*TritonServerApiVersionFn_t)(